requirement.txt
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
decorator==4.4.2
enum34==1.1.10
GKextCPy==0.4.1
graphkernels==0.2.1
ipython==5.10.0
ipython-genutils==0.2.0
numpy==1.16.2
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
pkgconfig==1.5.1
prompt-toolkit==1.0.18
ptyprocess==0.6.0
Pygments==2.5.2
python-igraph==0.8.2
scandir==1.10.0
simplegeneric==0.8.1
six==1.15.0
texttable==1.6.2
traitlets==4.3.3
wcwidth==0.2.5
